% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{apply_imputation_method}
\alias{apply_imputation_method}
\title{Apply imputation method to a numeric vector}
\usage{
apply_imputation_method(x, method, maxgap)
}
\arguments{
\item{x}{Numeric vector}

\item{method}{Imputation method}

\item{maxgap}{Maximum gap size}
}
\value{
Imputed vector
}
\description{
Apply imputation method to a numeric vector
}
\keyword{internal}
