% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ar_sat.R
\name{calc_Arsat}
\alias{calc_Arsat}
\title{Calculate dissolved argon saturation (deprecated)}
\usage{
calc_Arsat(
  temp_water,
  atmo_press,
  units = "atm",
  salinity = 0,
  out_units = "umol/L"
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=calc_ar_sat]{calc_ar_sat()}} instead.
}
\examples{
# Old:
# calc_Arsat(temp_water = 20, atmo_press = 1, salinity = 0)
# New:
calc_ar_sat(temp_water = 20, atmo_press = 1, salinity = 0)

}
\keyword{internal}
