% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ch4_sat.R
\name{calc_CH4sat}
\alias{calc_CH4sat}
\title{Calculate dissolved methane saturation (deprecated)}
\usage{
calc_CH4sat(
  temp_water,
  atmo_press,
  units = "atm",
  salinity = 0,
  xCH4_ppm = 1.9,
  out_units = "umol/L"
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=calc_ch4_sat]{calc_ch4_sat()}} instead.
}
\examples{
# Old:
# calc_CH4sat(temp_water = 20, atmo_press = 1, salinity = 0, xCH4_ppm = 1.9)
# New:
calc_ch4_sat(temp_water = 20, atmo_press = 1, salinity = 0, xch4_ppm = 1.9)

}
\keyword{internal}
