% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CO2_conc.R
\name{calc_CO2_molKg}
\alias{calc_CO2_molKg}
\title{Calculate dissolved CO2 concentration in mol/kg (deprecated)}
\usage{
calc_CO2_molKg(
  CO2_ppm,
  temp_water,
  water_depth_m,
  atmo_press,
  press_units,
  salinity = 0,
  waterDepth_m = lifecycle::deprecated()
)
}
\arguments{
\item{CO2_ppm}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{co2_ppm} instead.}

\item{waterDepth_m}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{water_depth_m}
instead.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=calc_co2_mol_kg]{calc_co2_mol_kg()}} instead.
}
\examples{
# Old:
# calc_CO2_molKg(
#   CO2_ppm = 420,
#   temp_water = 20,
#   water_depth_m = 0.5,
#   atmo_press = 101.325,
#   press_units = "kPa"
# )
# New:
calc_co2_mol_kg(
  co2_ppm = 420,
  temp_water = 20,
  water_depth_m = 0.5,
  atmo_press = 101.325,
  press_units = "kPa"
)

}
\keyword{internal}
