% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamMetabolizer.R
\name{calc_hour_angle}
\alias{calc_hour_angle}
\title{Calculate hour angle as in
http://education.gsfc.nasa.gov/experimental/July61999siteupdate/inv99Project.Site/Pages/solar.insolation.html.}
\usage{
calc_hour_angle(hour, format = c("degrees", "radians"))
}
\arguments{
\item{hour}{numeric value or vector. hour since (solar) midnight as number
between 0 and 23.999}

\item{format}{The format of both the input and the output. May be "degrees"
or "radians".}
}
\value{
numeric value or vector, in the units specified by \code{format},
indicating the angle corresponding to each value supplied in \code{hour}.
}
\description{
Adapted from \code{streamMetabolizer::calc_hour_angle} under CC0 license.
}
\details{
This is an approximation when hour is in clock time; should actually be given
in solar time
}
