% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_k0.R
\name{calc_K0}
\alias{calc_K0}
\title{Calculate the CO2 solubility coefficient (deprecated)}
\usage{
calc_K0(
  temp_water,
  water_depth_m = 0,
  atmo_press = 1,
  salinity = 0,
  waterDepth_m = lifecycle::deprecated()
)
}
\arguments{
\item{waterDepth_m}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{water_depth_m}
instead.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=calc_k0]{calc_k0()}} instead.
}
\examples{
# Old:
# calc_K0(temp_water = 20)
# New:
calc_k0(temp_water = 20)

}
\keyword{internal}
