% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_n2o_sat.R
\name{calc_N2Osat}
\alias{calc_N2Osat}
\title{Calculate dissolved nitrous oxide saturation (deprecated)}
\usage{
calc_N2Osat(
  temp_water,
  atmo_press,
  units = "atm",
  salinity = 0,
  xN2O_ppm = 0.338,
  out_units = "umol/L"
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=calc_n2o_sat]{calc_n2o_sat()}} instead.
}
\examples{
# Old:
# calc_N2Osat(temp_water = 20, atmo_press = 1, salinity = 0, xN2O_ppm = 0.338)
# New:
calc_n2o_sat(temp_water = 20, atmo_press = 1, salinity = 0, xn2o_ppm = 0.338)

}
\keyword{internal}
