% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_o2_sat.R
\name{calc_O2sat}
\alias{calc_O2sat}
\title{Calculate dissolved oxygen saturation (deprecated)}
\usage{
calc_O2sat(
  temp_water,
  atmo_press,
  units = "atm",
  salinity = 0,
  out_units = "mg/L"
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=calc_o2_sat]{calc_o2_sat()}} instead.
}
\examples{
# Old:
# calc_O2sat(temp_water = 15, atmo_press = 1, units = "atm")
# New:
calc_o2_sat(temp_water = 15, atmo_press = 1, units = "atm")

}
\keyword{internal}
