% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamMetabolizer.R
\name{calc_zenith_angle}
\alias{calc_zenith_angle}
\title{Calculate zenith angle as in
http://education.gsfc.nasa.gov/experimental/July61999siteupdate/inv99Project.Site/Pages/solar.insolation.html}
\usage{
calc_zenith_angle(
  latitude,
  declination.angle,
  hour.angle,
  format = c("degrees", "radians")
)
}
\arguments{
\item{latitude}{numeric value or vector indicating the site latitude in
decimal degrees (never radians or deg-min-sec, no matter what \code{format}
is) between -90 (South Pole) and 90 (North Pole).}

\item{declination.angle}{numeric value or vector, in the units specified by
\code{format}, indicating the declination angle.}

\item{hour.angle}{numeric value or vector, in the units specified by
\code{format}, indicating the angle.}

\item{format}{The format of both the output. May be "degrees" or "radians".}
}
\description{
Adapted from \code{streamMetabolizer::calc_hour_angle} under CC0 license.
}
