% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_par}
\alias{calculate_par}
\title{Calculate PAR with streamMetabolizer}
\usage{
calculate_par(df, site_meta)
}
\arguments{
\item{df}{Dataframe containing datetime data}

\item{site_meta}{Site metadata}
}
\value{
Dataframe with calculated solar metrics
}
\description{
Calculate PAR with streamMetabolizer
}
\keyword{internal}
