% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.R
\name{cdo_datacategories}
\alias{cdo_datacategories}
\title{Query the NCEI CDO data categories endpoint (deprecated)}
\usage{
cdo_datacategories(
  id = NULL,
  datasetid = NULL,
  locationid = NULL,
  stationid = NULL,
  startdate = NULL,
  enddate = NULL,
  sortfield = NULL,
  sortorder = NULL,
  max_results = Inf
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=cdo_data_categories]{cdo_data_categories()}} instead.
}
\examples{
# Old:
# cdo_datacategories(datasetid = "GHCND")
# New:
\dontrun{
cdo_data_categories(dataset_id = "GHCND")
}

}
\keyword{internal}
