% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.R
\name{cdo_datatypes}
\alias{cdo_datatypes}
\title{Query the NCEI CDO data types endpoint (deprecated)}
\usage{
cdo_datatypes(
  id = NULL,
  datasetid = NULL,
  locationid = NULL,
  stationid = NULL,
  datacategoryid = NULL,
  startdate = NULL,
  enddate = NULL,
  sortfield = NULL,
  sortorder = NULL,
  max_results = Inf
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=cdo_data_types]{cdo_data_types()}} instead.
}
\examples{
# Old:
# cdo_datatypes(datasetid = "GHCND", datacategoryid = "TEMP")
# New:
\dontrun{
cdo_data_types(dataset_id = "GHCND", data_category_id = "TEMP")
}

}
\keyword{internal}
