% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.R
\name{cdo_locationcategories}
\alias{cdo_locationcategories}
\title{Query the NCEI CDO location categories endpoint (deprecated)}
\usage{
cdo_locationcategories(
  id = NULL,
  datasetid = NULL,
  startdate = NULL,
  enddate = NULL,
  sortfield = NULL,
  sortorder = NULL,
  max_results = Inf
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=cdo_location_categories]{cdo_location_categories()}}
instead.
}
\examples{
# Old:
# cdo_locationcategories("ST")
# New:
\dontrun{
cdo_location_categories("ST")
}

}
\keyword{internal}
