% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{cfs_to_cms}
\alias{cfs_to_cms}
\title{Converts stream discharge in cubic feet per second to cubic meters per second}
\usage{
cfs_to_cms(Qcfs)
}
\arguments{
\item{Qcfs}{a numeric vector of stream discharge in cubic feet per second}
}
\value{
a numeric vector of stream discharge in cubic meters per second
}
\description{
Converts stream discharge in cubic feet per second to cubic meters per second
}
\examples{
Q <- seq(1,10,2)
cfs_to_cms(Q)
}
