% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_par}
\alias{clean_par}
\title{Clean and merge PAR values}
\usage{
clean_par(df)
}
\arguments{
\item{df}{Dataframe containing PAR data}
}
\value{
Cleaned dataframe
}
\description{
Clean and merge PAR values
}
\keyword{internal}
