% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_outliers.R
\name{consolidate_outlier_flags}
\alias{consolidate_outlier_flags}
\title{Consolidate Outlier Flags}
\usage{
consolidate_outlier_flags(df)
}
\arguments{
\item{df}{A dataframe containing \verb{_zFlag} columns.}
}
\value{
A dataframe with a consolidated \code{Flag} column.
}
\description{
Combines all zFlag columns into a single \code{Flag} column for easier interpretation.
}
