% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cfs.R
\name{convert_cfs}
\alias{convert_cfs}
\title{Convert Stream Discharge from Cubic Feet per Second to Various Units}
\usage{
convert_cfs(cfs, to = "cms")
}
\arguments{
\item{cfs}{A numeric vector of stream discharge in cubic feet per second.}

\item{to}{A character string specifying the target unit. Acceptable values are
"cms" (cubic meters per second) or "lps" (liters per second). Default is "cms".}
}
\value{
A numeric vector of stream discharge in the specified unit.
}
\description{
This function converts stream discharge measurements from cubic feet per second (cfs)
to cubic meters per second (cms) and liters per second (l/s).
}
\examples{
convert_cfs(cfs = 14.32, to = "cms")  # Converts 14.32 cfs to cms
convert_cfs(cfs = 14.32, to = "lps")  # Converts 14.32 cfs to lps
convert_cfs(cfs = c(10, 20, 30), to = "cms")  # Converts multiple values to cms
}
