% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamMetabolizer.R
\name{convert_PAR_to_SW}
\alias{convert_PAR_to_SW}
\title{Convert from photosynthetically active to shortwave radiation}
\usage{
convert_PAR_to_SW(par, coef = 0.473)
}
\arguments{
\item{par}{Vector of photosynthetically active radiation (400-700 nm;
umol/m^2/sec)}

\item{coef}{Numerical coefficient to convert PAR (umol/m^2/sec) to SW
(W/m^2). Defaults to value from Britton and Dodd (1976).}
}
\value{
Numeric vector of shortwave values with units W/m^2
}
\description{
Convert photosynthetically active radiation (PAR) to shortwave radiation
(SW). Uses a fixed ratio between PAR and SW, ignoring the minor seasonal
changes in this ratio (see Britton and Dodd (1976)).
}
\details{
This function was adapted from \code{streamMetabolizer::convert_PAR_to_SW},
which wraps a base conversion from \code{LakeMetabolizer::par.to.sw.base}. It is
included under the terms of the CC0 1.0 Universal public domain dedication:
https://www.usgs.gov/information-policies-and-instructions/copyrights-and-credits

Original citation:
Appling, A. P., Hall, R. O., Yackulic, C. B., & Arroita, M. (2018).
Overcoming Equifinality: Leveraging Long Time Series for Stream Metabolism
Estimation. \emph{Journal of Geophysical Research: Biogeosciences}, 123(2), 624–645.
https://doi.org/10.1002/2017JG004140
}
\examples{
convert_PAR_to_SW(par=400, coef=0.47)
}
