% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_pressure_to_atm}
\alias{convert_pressure_to_atm}
\title{Convert barometric pressure to atmospheres}
\usage{
convert_pressure_to_atm(pressure, units)
}
\arguments{
\item{pressure}{Numeric. Barometric pressure value(s) to be converted.}

\item{units}{Character. Units of the input barometric pressure. Accepted values are
\code{"atm"}, \code{"hPa"}, \code{"mbar"}, \code{"kPa"}, \code{"Torr"}, \code{"psi"}, and \code{"bar"}.}
}
\value{
Numeric. Barometric pressure in atmospheres ([atm]).
}
\description{
This helper function converts barometric pressure from various units
to atmospheres ([atm]). It is intended for internal use within the package.
}
\examples{
convert_pressure_to_atm(101.325, units = "kPa")
convert_pressure_to_atm(1013.25, units = "hPa")

}
