% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamMetabolizer.R
\name{convert_SW_to_PAR}
\alias{convert_SW_to_PAR}
\title{Convert from shortwave to photosynthetically active radiation}
\usage{
convert_SW_to_PAR(sw, coef = 2.114)
}
\arguments{
\item{sw}{Vector of shortwave radiation (W/m^2)}

\item{coef}{Numerical coefficient to convert SW (W/m^2) to PAR
(umol/m^2/sec). Defaults to value from Britton and Dodd (1976).}
}
\value{
Numeric vector of PAR values in units umol/m^2/sec
}
\description{
Convert shortwave radiation (SW) to photosynthetically active radiation
(PAR). Uses a fixed ratio between SW and PAR, ignoring the minor seasonal
changes in this ratio (see Britton and Dodd (1976)).
}
\details{
This function was adapted from \code{streamMetabolizer::convert_SW_to_PAR},
which wraps \code{LakeMetabolizer::sw.to.par.base}. It is included under the terms
of the CC0 1.0 Universal public domain dedication:
https://www.usgs.gov/information-policies-and-instructions/copyrights-and-credits

Original citation:
Appling, A. P., Hall, R. O., Yackulic, C. B., & Arroita, M. (2018).
Overcoming Equifinality: Leveraging Long Time Series for Stream Metabolism
Estimation. \emph{Journal of Geophysical Research: Biogeosciences}, 123(2), 624–645.
https://doi.org/10.1002/2017JG004140
}
\examples{
convert_SW_to_PAR(sw=800)
convert_SW_to_PAR(sw=800, coef=2.1)
}
