% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa_data.R
\name{download_site_data}
\alias{download_site_data}
\title{Download data for all sites in metadata}
\usage{
download_site_data(metadata)
}
\arguments{
\item{metadata}{Dataframe with site information}
}
\value{
Combined dataframe with all sites' data
}
\description{
Download data for all sites in metadata
}
\keyword{internal}
