% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa_data.R
\name{download_site_data_parallel}
\alias{download_site_data_parallel}
\title{Download data for all sites in metadata using parallelization}
\usage{
download_site_data_parallel(metadata)
}
\arguments{
\item{metadata}{Dataframe with site information}
}
\value{
Combined dataframe with all sites' data
}
\description{
Download data for all sites in metadata using parallelization
}
\keyword{internal}
