% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa_data.R
\name{fetch_single_site}
\alias{fetch_single_site}
\title{Download NASA POWER data for a single site}
\usage{
fetch_single_site(site, long, lat, start, end, elev)
}
\arguments{
\item{site}{Site identifier}

\item{long}{Longitude}

\item{lat}{Latitude}

\item{start}{Start date}

\item{end}{End date}

\item{elev}{Elevation in meters}
}
\value{
A dataframe with NASA POWER data
}
\description{
Download NASA POWER data for a single site
}
\keyword{internal}
