% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even_timesteps.R
\name{find_mode}
\alias{find_mode}
\title{Find the mode (most frequent value) in a vector}
\usage{
find_mode(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
The most frequent value(s) in the vector
}
\description{
Find the mode (most frequent value) in a vector
}
\keyword{internal}
