% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gap_fill_par}
\alias{gap_fill_par}
\title{Gap Fill PAR (light) data}
\usage{
gap_fill_par(df, site_meta)
}
\arguments{
\item{df}{Dataframe containing PAR data}

\item{site_meta}{Site metadata}
}
\value{
Dataframe with gaps in PAR values filled with calc_light() from streamMetabolizer
}
\description{
Gap Fill PAR (light) data
}
\keyword{internal}
