% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_remote_mtime}
\alias{get_remote_mtime}
\title{Get the Last Modified Time of a Remote File (Internal)}
\usage{
get_remote_mtime(url)
}
\arguments{
\item{url}{A character string specifying the URL of the remote file.}
}
\value{
A POSIXlt object representing the last modified timestamp of the remote file, or \code{NULL} if the information is unavailable or an error occurs.
}
\description{
This internal function retrieves the "Last-Modified" timestamp of a remote file by sending an HTTP HEAD request to the given URL. The result is memoised to cache results for repeated calls with the same URL.
}
\details{
The function sends an HTTP HEAD request to the specified URL using the \pkg{httr} package. If the server responds with a 200 status code and includes a "Last-Modified" header, the timestamp is parsed and returned. If the request fails or the "Last-Modified" header is missing, \code{NULL} is returned.
}
\note{
The function uses \pkg{memoise} to cache results, so repeated calls with the same URL will not trigger additional HTTP requests.
}
\keyword{internal}
