% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_site_metadata}
\alias{get_site_metadata}
\title{Get site-specific metadata}
\usage{
get_site_metadata(site_name, metadata)
}
\arguments{
\item{site_name}{Name of the site}

\item{metadata}{Full metadata dataframe}
}
\value{
Site-specific metadata
}
\description{
Get site-specific metadata
}
\keyword{internal}
