% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iem_networks.R
\name{iem_networks}
\alias{iem_networks}
\title{Get Iowa Environmental Mesonet network identifiers}
\usage{
iem_networks()
}
\value{
A tibble containing IEM network identifiers, network names, time
zones, geographic extents, and windrose update timestamps when available.
}
\description{
Retrieves the network table from the Iowa Environmental Mesonet (IEM) API.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The Iowa Environmental Mesonet groups stations into networks such as
\code{IA_ASOS}, \code{IA_COOP}, and other state, roadway, hydrological, and special
observing networks. Use \code{\link[=iem_networks]{iem_networks()}} to discover valid \code{network} values
for other IEM helpers.
}
\examples{
\dontrun{
networks <- iem_networks()
subset(networks, id == "IA_ASOS")
}

}
\references{
Iowa Environmental Mesonet API:
\url{https://mesonet.agron.iastate.edu/api/}

IEM API v1 documentation:
\url{https://mesonet.agron.iastate.edu/api/1/docs}
}
