% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iem_ob_history.R
\name{iem_obhistory}
\alias{iem_obhistory}
\title{Get one day of Iowa Environmental Mesonet observations (deprecated)}
\usage{
iem_obhistory(station, network, date = Sys.Date(), full = FALSE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=iem_ob_history]{iem_ob_history()}} instead.
}
\examples{
# Old:
# iem_obhistory("DSM", network = "IA_ASOS", date = "2024-06-01")
# New:
\dontrun{
iem_ob_history("DSM", network = "IA_ASOS", date = "2024-06-01")
}

}
\keyword{internal}
