% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_numeric_column}
\alias{impute_numeric_column}
\title{Impute a single numeric column}
\usage{
impute_numeric_column(df, col, method, maxgap)
}
\arguments{
\item{df}{Dataframe containing the column}

\item{col}{Column name}

\item{method}{Imputation method}

\item{maxgap}{Maximum gap size to fill}
}
\value{
Dataframe with imputed column
}
\description{
Impute a single numeric column
}
\keyword{internal}
