% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_numeric_columns}
\alias{impute_numeric_columns}
\title{Impute numeric columns using specified method}
\usage{
impute_numeric_columns(df, method, maxgap)
}
\arguments{
\item{df}{Dataframe containing numeric columns}

\item{method}{Imputation method}

\item{maxgap}{Maximum gap size to fill}
}
\value{
Dataframe with imputed numeric columns
}
\description{
Impute numeric columns using specified method
}
\keyword{internal}
