% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_single_site}
\alias{impute_single_site}
\title{Impute values for a single site}
\usage{
impute_single_site(site_df, metadata, method, maxgap)
}
\arguments{
\item{site_df}{Single site dataframe}

\item{metadata}{Site metadata}

\item{method}{Imputation method}

\item{maxgap}{Maximum gap size to fill}
}
\value{
Imputed dataframe for single site
}
\description{
Impute values for a single site
}
\keyword{internal}
