% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kings_discharge}
\alias{kings_discharge}
\title{Kings Creek daily water data}
\format{
A tibble with 716 rows and 5 variables:
\describe{
\item{monitoring_location_id}{USGS monitoring location ID.}
\item{time}{Date of observation.}
\item{value}{Observed value. Units depend on \code{parameter_code}.}
\item{parameter_code}{USGS parameter code: \code{00060} = discharge,
\code{00065} = gage height, \code{00010} = water temperature.}
\item{qualifier}{USGS data qualifier.}
}
}
\source{
USGS Water Data API, retrieved with \code{dataRetrieval::read_waterdata_daily()}.
}
\usage{
kings_discharge
}
\description{
Daily USGS water data for Kings Creek at monitoring location
\code{USGS-06879650}.
}
\details{
The dataset includes daily observations for water year 2025,
from 2024-10-01 through 2025-09-30.
}
\keyword{datasets}
