% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_mesonet.R
\name{ks_meso_fw13}
\alias{ks_meso_fw13}
\title{Get Kansas Mesonet FW13 Data}
\usage{
ks_meso_fw13(station, start_date, end_date)
}
\arguments{
\item{station}{Station name as a character string.}

\item{start_date}{Start date for the data retrieval in \code{YYYY-MM-DD} or
\code{YYYYMMDD} format.}

\item{end_date}{End date for the data retrieval in \code{YYYY-MM-DD} or \code{YYYYMMDD}
format.}
}
\value{
A character vector containing FW13 records.
}
\description{
Retrieves fire weather data in FW13 format for one station and date range.
}
