% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_mesonet.R
\name{ks_meso_most_recent}
\alias{ks_meso_most_recent}
\title{Get Most Recent Kansas Mesonet Data Timestamp}
\usage{
ks_meso_most_recent(interval)
}
\arguments{
\item{interval}{Data interval. Must be one of \code{"hour"}, \code{"5min"}, or
\code{"day"}.}
}
\value{
A data frame with station names and most recent observation times.
}
\description{
Retrieves the timestamp of the most recently ingested data for each station
for a given observation interval. The returned data includes the \code{"--all--"}
row reported by Kansas Mesonet.
}
