% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_meso_stations.R
\name{ks_meso_stations}
\alias{ks_meso_stations}
\title{Get Kansas Mesonet station information}
\usage{
ks_meso_stations()
}
\value{
A data frame containing station metadata.
}
\description{
Fetches metadata about Kansas Mesonet stations, including location and
network details.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Kansas Mesonet data are preliminary and subject to revision. Cite the Kansas
Mesonet when sharing, publishing, or otherwise disseminating data accessed
with this function. A suggested citation format is: Kansas Mesonet, year:
webpage title. Accessed date, webpage URL. Review the Kansas Mesonet data
usage policy before automated use; automated page scraping or data ingesting
without written consent is not permitted.
}
\examples{
\dontrun{
stations <- ks_meso_stations()
subset(stations, grepl("Konza", StationName))
}

}
\references{
Kansas Mesonet data usage policy:
\url{https://mesonet.k-state.edu/about/usage/}
}
