% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_meso_time_series.R
\name{ks_meso_timeseries}
\alias{ks_meso_timeseries}
\title{Get Kansas Mesonet time-series data (deprecated)}
\usage{
ks_meso_timeseries(
  stations = NULL,
  network = NULL,
  start_date,
  end_date,
  interval,
  vars = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=ks_meso_time_series]{ks_meso_time_series()}} instead.
}
\examples{
# Old:
# ks_meso_timeseries(stations = "Konza Prairie", start_date = "2024-06-01",
#   end_date = "2024-06-07", interval = "hour")
# New:
\dontrun{
ks_meso_time_series(
  stations = "Konza Prairie",
  start_date = "2024-06-01",
  end_date = "2024-06-07",
  interval = "hour"
)
}

}
\keyword{internal}
