% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei_stations.R
\name{ncei_bbox}
\alias{ncei_bbox}
\title{Compute a bounding box around a point}
\usage{
ncei_bbox(latitude, longitude, dist_km)
}
\arguments{
\item{latitude, longitude}{Numeric. Target coordinates in decimal
degrees. Latitude must be between -90 and 90; longitude between
-180 and 180.}

\item{dist_km}{Positive numeric. Half-width of the bounding box in
kilometres.}
}
\value{
A numeric vector \code{c(north, west, south, east)}.
}
\description{
Returns a bounding box vector suitable for the \code{bbox} argument of
\code{\link[=ncei_stations]{ncei_stations()}} and \code{\link[=get_noaa_stations]{get_noaa_stations()}}.
}
\examples{
# Bounding box 50 km around Konza Prairie
ncei_bbox(39.1, -96.6, 50)

}
