% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xco2_to_pco2.R
\name{pCO2_to_xCO2}
\alias{pCO2_to_xCO2}
\title{Convert CO2 partial pressure to mole fraction (deprecated)}
\usage{
pCO2_to_xCO2(temp_water, pCO2_uatm, atmo_press, press_units, ...)
}
\arguments{
\item{pCO2_uatm}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{pco2_uatm}
instead.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=pco2_to_xco2]{pco2_to_xco2()}} instead.
}
\examples{
# Old:
# pCO2_to_xCO2(
#   temp_water = 20,
#   pCO2_uatm = 400,
#   atmo_press = 101.325,
#   press_units = "kPa"
# )
# New:
pco2_to_xco2(
  temp_water = 20,
  pco2_uatm = 400,
  atmo_press = 101.325,
  press_units = "kPa"
)

}
\keyword{internal}
