% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa_data.R
\name{process_nasa_data}
\alias{process_nasa_data}
\title{Process NASA POWER data}
\usage{
process_nasa_data(raw_data)
}
\arguments{
\item{raw_data}{Combined raw data from all sites}
}
\value{
Processed tibble with required columns
}
\description{
Process NASA POWER data
}
\keyword{internal}
