% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDO.R
\name{readDO}
\alias{readDO}
\title{Fast and friendly data logger file finagler}
\usage{
readDO(path, type = c("PME", "YSI", "AIMS"))
}
\arguments{
\item{path}{File name in working directory or path to logger files}

\item{type}{Type of data logger, currently supports PME miniDOT, YSI EXO, or the AIMS pipeline}
}
\value{
a data.table of logger measurements with nice column names
}
\description{
Load and concatenate files from dissolved oxygen loggers, may work for other parameters from supported loggers
}
