% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove Flagged Values}
\usage{
remove_outliers(df)
}
\arguments{
\item{df}{A dataframe containing a \code{Flag} column.}
}
\value{
A dataframe with flagged values replaced by NA.
}
\description{
Sets flagged values in the dataframe to NA based on the \code{Flag} column.
}
