% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_by_site}
\alias{split_by_site}
\title{Split dataframe by site}
\usage{
split_by_site(df)
}
\arguments{
\item{df}{Dataframe to split}
}
\value{
List of dataframes split by site
}
\description{
Split dataframe by site
}
\keyword{internal}
