% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_meso_time_series.R
\name{tex_meso_timeseries}
\alias{tex_meso_timeseries}
\title{Get recent TexMesonet time-series data (deprecated)}
\usage{
tex_meso_timeseries(
  site_id,
  prior_minutes,
  variable = c("all", "temperature", "humidity", "barometric_pressure", "precip",
    "wind_speed")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=tex_meso_time_series]{tex_meso_time_series()}} instead.
}
\examples{
# Old:
# tex_meso_timeseries(2, prior_minutes = 60)
# New:
tex_meso_time_series(2, prior_minutes = 60)

}
\keyword{internal}
