% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamMetabolizer.R
\name{to_degrees}
\alias{to_degrees}
\title{Convert radians to degrees}
\usage{
to_degrees(radians)
}
\arguments{
\item{radians}{angle in radians}
}
\value{
angle in degrees
}
\description{
Converts radians to degrees.
}
\details{
Adapted from \code{streamMetabolizer}, included under the CC0 1.0 Universal public domain dedication.
}
