% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamMetabolizer.R
\name{to_radians}
\alias{to_radians}
\title{Convert degrees to radians}
\usage{
to_radians(degrees)
}
\arguments{
\item{degrees}{angle in degrees}
}
\value{
angle in radians
}
\description{
Converts degrees to radians.
}
\details{
Adapted from \code{streamMetabolizer}, included under the CC0 1.0 Universal public domain dedication:
https://www.usgs.gov/information-policies-and-instructions/copyrights-and-credits
}
