% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Validate input data structure}
\usage{
validate_input_data(df, metadata)
}
\arguments{
\item{df}{Input dataframe}

\item{metadata}{Metadata dataframe}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate input data structure
}
\keyword{internal}
