% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate imputation parameters}
\usage{
validate_parameters(method, maxgap)
}
\arguments{
\item{method}{Imputation method}

\item{maxgap}{Maximum gap size}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate imputation parameters
}
\keyword{internal}
