% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xco2_to_pco2.R
\name{xCO2_to_pCO2}
\alias{xCO2_to_pCO2}
\title{Convert CO2 mole fraction to partial pressure (deprecated)}
\usage{
xCO2_to_pCO2(xCO2_ppm, temp_water, atmo_press, press_units, ...)
}
\arguments{
\item{xCO2_ppm}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{xco2_ppm} instead.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=xco2_to_pco2]{xco2_to_pco2()}} instead.
}
\examples{
# Old:
# xCO2_to_pCO2(
#   xCO2_ppm = 420,
#   temp_water = 20,
#   atmo_press = 101.325,
#   press_units = "kPa"
# )
# New:
xco2_to_pco2(
  xco2_ppm = 420,
  temp_water = 20,
  atmo_press = 101.325,
  press_units = "kPa"
)

}
\keyword{internal}
