% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CO2_conc.R
\name{calc_CO2_molKg}
\alias{calc_CO2_molKg}
\title{Calculate CO2 Concentration from Sensor Data}
\usage{
calc_CO2_molKg(
  CO2_ppm,
  temp_water,
  waterDepth_m,
  atmo_press,
  press_units,
  salinity = 0
)
}
\arguments{
\item{CO2_ppm}{Numeric. Mole fraction of CO2 in parts per million.}

\item{temp_water}{Numeric. Water temperature in degrees Celsius.}

\item{waterDepth_m}{Numeric. Water depth in meters.}

\item{atmo_press}{Numeric. Atmospheric pressure.}

\item{press_units}{Character. Units of atmospheric pressure.}

\item{salinity}{Numeric. Salinity in PSU.}
}
\value{
a numeric vector of CO2 concentration in mol/kg.
}
\description{
Calculate CO2 Concentration from Sensor Data
}
