% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_co2_sat.R
\name{calc_CO2sat}
\alias{calc_CO2sat}
\title{Calculate dissolved carbon dioxide saturation (deprecated)}
\usage{
calc_CO2sat(
  temp_water,
  atmo_press,
  units = "atm",
  salinity = 0,
  xCO2_ppm = 420,
  out_units = "umol/L"
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=calc_co2_sat]{calc_co2_sat()}} instead.
}
\examples{
# Old:
# calc_CO2sat(temp_water = 20, atmo_press = 1, salinity = 0, xCO2_ppm = 420)
# New:
calc_co2_sat(temp_water = 20, atmo_press = 1, salinity = 0, xco2_ppm = 420)

}
\keyword{internal}
