% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_K0.R
\name{calc_K0}
\alias{calc_K0}
\title{Calculate Solubility Constant (K0) for CO2 in Water}
\usage{
calc_K0(temp_water, waterDepth_m = 0, atmo_press = 1, salinity = 0)
}
\arguments{
\item{temp_water}{Numeric. Water temperature in degrees Celsius.}

\item{waterDepth_m}{Numeric. Water depth in meters.}

\item{atmo_press}{Numeric. Atmospheric pressure in atm.}

\item{salinity}{Numeric. Salinity in PSU (Practical Salinity Units).}
}
\value{
Numeric. Solubility constant (K0) in mol/(kg·atm).
}
\description{
This function calculates the solubility constant (K0) for CO2 in water
as a function of temperature and salinity, based on Weiss (1974).
}
\references{
Weiss, R.F. (1974). Carbon dioxide in water and seawater: the solubility of
a non-ideal gas. Marine Chemistry, 2, 203-215.
}
\keyword{internal}
