% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamMetabolizer.R
\name{calc_light}
\alias{calc_light}
\title{Calculate modeled light from solar.time}
\usage{
calc_light(solar.time, latitude, longitude, max.PAR = 2326)
}
\arguments{
\item{solar.time}{mean solar time, as required for input to metabolism
models.}

\item{latitude}{numeric value or vector indicating the site latitude in
decimal degrees (never radians or deg-min-sec, no matter what \code{format}
is) between -90 (South Pole) and 90 (North Pole).}

\item{longitude}{Numeric longitude in decimal degrees (east positive, west negative).}

\item{max.PAR}{numeric: the PAR (umol m^-2 s^-1) that each
day should reach at peak light}
}
\description{
Calculate photosynthetically active radiation (PAR) for a series of
date-times and site coordinates. This function was adapted from
\code{streamMetabolizer::calc_light} and is included here under the terms of the
CC0 1.0 Universal public domain dedication, as described at:
https://www.usgs.gov/information-policies-and-instructions/copyrights-and-credits
}
\details{
Original citation:
Appling, A. P., Hall, R. O., Yackulic, C. B., & Arroita, M. (2018).
Overcoming Equifinality: Leveraging Long Time Series for Stream Metabolism
Estimation. \emph{Journal of Geophysical Research: Biogeosciences}, 123(2), 624–645.
https://doi.org/10.1002/2017JG004140
}
\examples{
solar.time <- lubridate::force_tz(as.POSIXct('2016-09-27 12:00'), 'UTC')
calc_light(solar.time, 40, -120)
}
