% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_n2_sat.R
\name{calc_N2sat}
\alias{calc_N2sat}
\title{Calculate dissolved nitrogen saturation (deprecated)}
\usage{
calc_N2sat(
  temp_water,
  atmo_press,
  units = "atm",
  salinity = 0,
  out_units = "umol/L"
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{\link[=calc_n2_sat]{calc_n2_sat()}} instead.
}
\examples{
# Old:
# calc_N2sat(temp_water = 20, atmo_press = 1, salinity = 0)
# New:
calc_n2_sat(temp_water = 20, atmo_press = 1, salinity = 0)

}
\keyword{internal}
