% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_water_density.R
\name{calc_water_density}
\alias{calc_water_density}
\title{Calculate Water Density}
\usage{
calc_water_density(tempC)
}
\arguments{
\item{tempC}{Numeric value representing the water temperature in degrees Celsius.}
}
\value{
A numeric value or vector representing the density of water in kg/m^3.
}
\description{
This function calculates the density of water (kg/m^3) based on the temperature in degrees Celsius.
The equation is based on the formulation by Kell (1975) for pure water.
}
\details{
This function uses the polynomial approximation provided by Kell (1975) for the density of water
as a function of temperature. This approximation is valid for temperatures between 0°C and 150°C.
}
\examples{
# Example usage:
calc_water_density(tempC = 15)

}
\references{
Kell, G. S. (1975). Density, thermal expansivity, and compressibility of liquid water from 0° to 150°C:
Correlations and tables for atmospheric pressure and saturation reviewed and expressed on 1968 temperature scale.
\emph{Journal of Chemical and Engineering Data}, 20(1), 97–105.
\doi{10.1021/je60064a005}
}
