% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_flow.R
\name{convert_flow}
\alias{convert_flow}
\title{Convert Stream Discharge to Various Units}
\usage{
convert_flow(flow, to = "cms")
}
\arguments{
\item{flow}{A numeric vector or a \code{units} object of stream discharge. If numeric,
it is assumed to be in cubic feet per second.}

\item{to}{A character string specifying the target unit. Acceptable values are
"cfs" (cubic feet per second), "cms" (cubic meters per second), or "lps" (liters per second).
Default is "cms".}
}
\value{
A \code{units} object of stream discharge in the specified unit.
}
\description{
This function converts stream discharge measurements between cubic feet per second (cfs),
cubic meters per second (cms), and liters per second (lps). If the input does not have
units, it is assumed to be in cubic feet per second.
}
\examples{
convert_flow(flow = 14.32, to = "cms")  # Converts 14.32 cfs to cms
convert_flow(flow = 14.32, to = "lps")  # Converts 14.32 cfs to lps
convert_flow(flow = units::set_units(14.32, "ft^3/s"), to = "cms")  # With explicit units
convert_flow(flow = c(10, 20, 30), to = "cms")  # Converts multiple values from assumed cfs
}
