% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamMetabolizer.R
\name{convert_solartime_to_UTC}
\alias{convert_solartime_to_UTC}
\title{Convert DateTime from Local Solar Time to UTC}
\usage{
convert_solartime_to_UTC(
  any.solar.time,
  longitude,
  time.type = c("apparent solar", "mean solar")
)
}
\arguments{
\item{any.solar.time}{A POSIXct datetime object in solar time (mean or apparent). Must have tz = "UTC".}

\item{longitude}{Numeric longitude in decimal degrees (east positive, west negative).}

\item{time.type}{Character string: either \code{"mean solar"} or \code{"apparent solar"}.
"apparent solar", i.e. true solar time, is noon when the sun is at its zenith.
"mean solar" approximates apparent solar time but with noons exactly 24 hours apart.}
}
\value{
A POSIXct datetime in UTC.
}
\description{
Converts a datetime from local solar time back to UTC. Input time may be either
apparent solar time (sun at zenith at noon) or mean solar time (noons exactly 24 hours apart).
}
\references{
Yard, Bennett, Mietz, Coggins, Stevens, Hueftle, and Blinn. 2005.
Influence of topographic complexity on solar insolation estimates for the
Colorado River, Grand Canyon, AZ. Ecological Modelling.
}
