% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamMetabolizer.R
\name{convert_UTC_to_solartime}
\alias{convert_UTC_to_solartime}
\title{Convert UTC Time to Mean or Apparent Solar Time}
\usage{
convert_UTC_to_solartime(
  date.time,
  longitude,
  time.type = c("mean solar", "apparent solar")
)
}
\arguments{
\item{date.time}{A POSIXct datetime object in UTC timezone.}

\item{longitude}{Longitude in decimal degrees (east positive, west negative).}

\item{time.type}{One of \code{"mean solar"} (default) or \code{"apparent solar"}.}
}
\value{
A POSIXct datetime in mean or apparent solar time (still tz = "UTC").
}
\description{
Converts a UTC datetime to mean or apparent solar time for a given longitude.
Uses a high-precision offset for longitude and applies the equation of time
for apparent solar time.
}
\details{
"apparent solar", i.e. true solar time, is noon when the sun is at its zenith.
"mean solar" approximates apparent solar time but with noons exactly 24 hours apart.
Elsewhere in this package, variables named \code{solar.time} are mean solar time.
}
\examples{
utc <- as.POSIXct("2025-04-30 12:00:00", tz = "UTC")
convert_UTC_to_solartime(utc, -90, time.type = "mean solar")
convert_UTC_to_solartime(utc, -90, time.type = "apparent solar")

}
