% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_bp.R
\name{correct_bp}
\alias{correct_bp}
\title{Correct Barometric Pressure for Elevation Change}
\usage{
correct_bp(station_kPa, air_temp, station_elev, site_elev)
}
\arguments{
\item{station_kPa}{A vector of station barometric pressures in kPa.}

\item{air_temp}{A vector of air temperatures at the station in degrees Celsius.}

\item{station_elev}{Elevation of the initial site in meters.}

\item{site_elev}{Elevation of the target site in meters.}
}
\value{
A vector of corrected barometric pressures at the new elevation in kPa.
}
\description{
This function adjusts barometric pressure from one elevation to another based on the barometric formula.
}
\examples{
correct_bp(station_kPa = c(101.3, 100.5), air_temp = c(15, 10), station_elev = 300, site_elev = 500)

}
