% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTimesteps.R
\name{find_mode}
\alias{find_mode}
\title{Fast, easy, and even time steps}
\usage{
find_mode(loggerData)
}
\arguments{
\item{loggerData}{logger dataframe with a DateTime_UTC column}
}
\value{
a dataframe
}
\description{
easily get even and matching timesteps for a dataframe
}
