% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ks_meso.R
\name{get_ks_meso}
\alias{get_ks_meso}
\title{Fetch data from Kansas Mesonet}
\usage{
get_ks_meso(
  stations = NULL,
  network = NULL,
  start_date,
  end_date,
  interval,
  vars = NULL,
  output_dir = NULL,
  debug = TRUE
)
}
\arguments{
\item{stations}{Character vector of station names to retrieve data for. Use
\code{"all"} to retrieve data for all stations. Must be \code{NULL} when \code{network}
is supplied.}

\item{network}{Network name to retrieve data for, one of:
\itemize{
\item \code{"BBW"}: Big Bend Groundwater Management District
\item \code{"EBW"}: Equus Beds Groundwater Management District
\item \code{"KSRE"}: K-State Research and Extension
}

This is an alternative to \code{stations}.}

\item{start_date}{Start date for the data retrieval in \code{YYYY-MM-DD} format.}

\item{end_date}{End date for the data retrieval in \code{YYYY-MM-DD} format.}

\item{interval}{Data interval. Must be one of \code{"hour"}, \code{"5min"}, or
\code{"day"}.}

\item{vars}{Character vector of variables to retrieve. Defaults to common variables.}

\item{output_dir}{Directory to save the downloaded data. Defaults to the cache path.}

\item{debug}{Logical; if \code{TRUE}, debug messages are printed.}
}
\value{
A list with details about successful and failed downloads, output
directory, and data chunks.
}
\description{
Retrieves weather data for specified stations from the Kansas Mesonet.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Kansas Mesonet data are preliminary and subject to revision. Cite the Kansas
Mesonet when sharing, publishing, or otherwise disseminating data accessed
with this function. A suggested citation format is: Kansas Mesonet, year:
webpage title. Accessed date, webpage URL. Review the Kansas Mesonet data
usage policy before automated use; automated page scraping or data ingesting
without written consent is not permitted.
}
\examples{
\dontrun{
get_ks_meso(
  stations = "Konza Prairie",
  start_date = "2024-06-01",
  end_date = "2024-06-07",
  interval = "hour",
  vars = c("TEMP2MAVG", "RELHUM2MAVG", "PRESSUREAVG")
)
}

}
\references{
Kansas Mesonet data usage policy:
\url{https://mesonet.k-state.edu/about/usage/}
}
