% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_mesonet.R
\name{get_ks_meso}
\alias{get_ks_meso}
\title{Fetch Data from Kansas Mesonet}
\usage{
get_ks_meso(
  stations,
  start_date,
  end_date,
  interval,
  vars = NULL,
  output_dir = NULL,
  debug = T
)
}
\arguments{
\item{stations}{Character vector of station names to retrieve data for.}

\item{start_date}{Start date for the data retrieval in \code{YYYY-MM-DD} format.}

\item{end_date}{End date for the data retrieval in \code{YYYY-MM-DD} format.}

\item{interval}{Data interval. Must be one of \code{'hour'}, \code{'5min'}, or \code{'day'}.}

\item{vars}{Character vector of variables to retrieve. Defaults to common variables.}

\item{output_dir}{Directory to save the downloaded data. Defaults to the cache path.}

\item{debug}{Logical; if \code{TRUE}, debug messages are printed.}
}
\value{
A list with details about successful and failed downloads, output directory, and data chunks.
}
\description{
Retrieves weather data for specified stations from the Kansas Mesonet.
}
