% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa_data.R
\name{get_nasa_data}
\alias{get_nasa_data}
\title{Download NASA POWER Hourly Data}
\usage{
get_nasa_data(
  site_meta,
  params = c("PSC", "ALLSKY_SFC_SW_DWN", "PRECTOTCORR", "T2M"),
  max_attempts = 5
)
}
\arguments{
\item{site_meta}{A data frame containing metadata for each site. Required columns:
\itemize{
\item \code{Site}: site name (character)
\item \code{lat}: latitude (numeric)
\item \code{lon}: longitude (numeric)
\item \code{elev_m}: elevation in meters (numeric)
\item \code{start_date}, \code{end_date}: date range (can be \code{Date} or character)
}}

\item{params}{Case-insensitive character vector of solar, meteorological or climatology parameters to download. Defaults to "PSC", "ALLSKY_SFC_SW_DWN", "PRECTOTCORR", and "T2M". See \link[nasapower]{get_power} for more information.}

\item{max_attempts}{Number of retry attempts for failed API calls. Default is 5.}
}
\value{
A tibble with columns:
\describe{
\item{Site}{Site name}
\item{dateTime}{timestamp in UTC}
\item{PSC}{Elevation-corrected barometric pressure (kPa)}
\item{ALLSKY_SFC_SW_DWN}{All Sky Surface Shortwave Downward Irradiance (W/m²)}
\item{T2M}{Average air temperature at 2 m above the surface (°C)}
\item{PRECTOTCORR}{MERRA-2 bias corrected total precipitation(mm/hr)}
}
}
\description{
Downloads hourly meteorological data from the NASA POWER project for a set of sites
over specified time periods.
}
