% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season}
\alias{get_season}
\title{Determine the Season from a Date}
\usage{
get_season(date)
}
\arguments{
\item{date}{A character string or Date object. If a character string, it should be in a format that can be coerced to a Date.}
}
\value{
A character string indicating the season: "Winter", "Spring", "Summer", or "Fall".
}
\description{
This function determines the meteorological season (Winter, Spring, Summer, or Fall) based on the input date.
It handles invalid inputs and provides clear error messages.
}
\examples{
get_season("2024-12-25")
get_season(as.Date("2024-07-04"))
}
