% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_meso_fw13.R
\name{ks_meso_fw13}
\alias{ks_meso_fw13}
\title{Get Kansas Mesonet FW13 Data}
\usage{
ks_meso_fw13(station, start_date, end_date)
}
\arguments{
\item{station}{Station name as a character string.}

\item{start_date}{Start date for the data retrieval in \code{YYYY-MM-DD} or
\code{YYYYMMDD} format.}

\item{end_date}{End date for the data retrieval in \code{YYYY-MM-DD} or \code{YYYYMMDD}
format.}
}
\value{
A character vector containing FW13 records.
}
\description{
Retrieves fire weather data in FW13 format for one station and date range.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Kansas Mesonet data are preliminary and subject to revision. Cite the Kansas
Mesonet when sharing, publishing, or otherwise disseminating data accessed
with this function. A suggested citation format is: Kansas Mesonet, year:
webpage title. Accessed date, webpage URL. Review the Kansas Mesonet data
usage policy before automated use; automated page scraping or data ingesting
without written consent is not permitted.
}
\references{
Kansas Mesonet data usage policy:
\url{https://mesonet.k-state.edu/about/usage/}
}
