% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_meso_most_recent.R
\name{ks_meso_most_recent}
\alias{ks_meso_most_recent}
\title{Get most recent Kansas Mesonet data timestamp}
\usage{
ks_meso_most_recent(interval)
}
\arguments{
\item{interval}{Data interval. Must be one of \code{"hour"}, \code{"5min"}, or
\code{"day"}.}
}
\value{
A data frame with station names and most recent observation times.
}
\description{
Retrieves the timestamp of the most recently ingested data for each station
for a given observation interval. The returned data includes the \code{"--all--"}
row reported by Kansas Mesonet.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Kansas Mesonet data are preliminary and subject to revision. Cite the Kansas
Mesonet when sharing, publishing, or otherwise disseminating data accessed
with this function. A suggested citation format is: Kansas Mesonet, year:
webpage title. Accessed date, webpage URL. Review the Kansas Mesonet data
usage policy before automated use; automated page scraping or data ingesting
without written consent is not permitted.
}
\examples{
\dontrun{
ks_meso_most_recent(interval = "hour")
}

}
\references{
Kansas Mesonet data usage policy:
\url{https://mesonet.k-state.edu/about/usage/}
}
