% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_mesonet.R
\name{ks_meso_station_activity}
\alias{ks_meso_station_activity}
\title{Get Kansas Mesonet Station Activity}
\usage{
ks_meso_station_activity()
}
\value{
A data frame with station activity details, including start and end observation times.
}
\description{
Retrieves activity data for Kansas Mesonet stations, including observation intervals and data spans.
}
