% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCO2_to_pCO2.R
\name{pCO2_to_xCO2}
\alias{pCO2_to_xCO2}
\title{Convert p(CO2) to x(CO2)}
\usage{
pCO2_to_xCO2(temp_water, pCO2_uatm, atmo_press, press_units, ...)
}
\arguments{
\item{temp_water}{Water temperature in degrees Celsius.}

\item{pCO2_uatm}{Partial pressure of CO2 in µatm.}

\item{atmo_press}{Atmospheric pressure.}

\item{press_units}{Units for pressure, either "atm" or others (default: "atm").}

\item{...}{Additional parameters passed to \code{calc_vapor_press}.}
}
\value{
CO2 mole fraction (xCO2) in ppm.
}
\description{
Converts partial pressure of CO2 p(CO2) µatm to mole fraction of CO2 x(CO2) ppm. This involves corrections for water vapor pressure to account for equilibrium conditions.
}
\references{
Dickson, A.G., Sabine, C.L., and Christian, J.R. (Eds.) (2007). Guide to best practices for
ocean CO2 measurements. PICES Special Publication 3, 191 pp.
}
\keyword{internal}
