% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_meso_timeseries.R
\name{tex_meso_timeseries}
\alias{tex_meso_timeseries}
\title{Get recent TexMesonet time-series data}
\usage{
tex_meso_timeseries(
  site_id,
  prior_minutes,
  variable = c("all", "temperature", "humidity", "barometric_pressure", "precip",
    "wind_speed")
)
}
\arguments{
\item{site_id}{TexMesonet station ID.}

\item{prior_minutes}{Number of minutes before the current time to retrieve.}

\item{variable}{Data type to retrieve. Use \code{"all"} for the charting fields
endpoint, or one of \code{"temperature"}, \code{"humidity"},
\code{"barometric_pressure"}, \code{"precip"}, or \code{"wind_speed"} for the
single-variable endpoints.}
}
\value{
A tibble of observations with UTC \code{dateTime} values. Single-variable
requests return \code{value} and \code{dateTime} columns and have \code{"field_name"},
\code{"station_name"}, \code{"twdb_station_id"}, and \code{"units"} attributes. Requests
with \code{variable = "all"} return one column per charting field and have
\code{"station_name"} and \code{"twdb_station_id"} attributes.
}
\description{
Retrieves recent time-series observations for a single TWDB TexMesonet
station.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

TexMesonet's public time-series API retrieves recent observations by station
and look-back window. The API may return different fields by station because
not all stations measure every parameter.
}
\examples{
\dontrun{
tex_meso_timeseries(2, prior_minutes = 60)
tex_meso_timeseries(2, prior_minutes = 60, variable = "temperature")
}

}
\references{
TexMesonet APIs:
\url{https://www.texmesonet.org/Apis}
}
