% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCO2_to_pCO2.R
\name{xCO2_to_pCO2}
\alias{xCO2_to_pCO2}
\title{Convert Mole Fraction of CO2 (xCO2) to Partial Pressure (pCO2)}
\usage{
xCO2_to_pCO2(xCO2_ppm, temp_water, atmo_press, press_units, ...)
}
\arguments{
\item{xCO2_ppm}{Numeric. Mole fraction of CO2 in air in parts per million (ppm). Default is 400 ppm.}

\item{temp_water}{Numeric. Water temperature in degrees Celsius.}

\item{atmo_press}{Numeric. Atmospheric pressure.}

\item{press_units}{Character. Units of atmospheric pressure.}

\item{...}{Additional arguments passed to the \code{calc_vapor_press} function.}
}
\value{
Numeric. Partial pressure of CO2 in µatm.
}
\description{
This function calculates the partial pressure of carbon dioxide (pCO2) in water
from its mole fraction (xCO2) in air, considering temperature and atmospheric pressure.
}
\details{
The partial pressure of CO2 (\eqn{pCO2}) is computed using the equation:
\deqn{pCO2 = (P_{atm} - P_{H2O}) \cdot xCO2}
where \eqn{P_{atm}} is the atmospheric pressure, \eqn{P_{H2O}} is the water vapor pressure,
and \eqn{xCO2} is the mole fraction of CO2 in air.
}
\references{
\itemize{
\item Dickson, A.G., Sabine, C.L., & Christian, J.R. (2007). \emph{Guide to Best Practices for Ocean CO2 Measurements}.
\item Weiss, R.F. (1974). Carbon dioxide in water and seawater: the solubility of a non-ideal gas. Marine Chemistry, 2(3), 203–215.
}
}
\keyword{internal}
